# Conflicts:
#	isa2/demo/src/main/java/isa2/demo/Controller/EntityController.java
#	isa2/demo/src/main/resources/data.sql

+++ isa2/demo/src/main/java/isa2/demo/Service/ServiceImpl/AdventureServiceImpl.java
@@ -1,18 +1,35 @@
 package isa2.demo.Service.ServiceImpl;
 
+import isa2.demo.DTO.FreeEntityDTO;
+import isa2.demo.Exception.InvalidInputException;
 import isa2.demo.Model.Adventure;
+import isa2.demo.Model.Owner;
+import isa2.demo.Model.Reservation;
+import isa2.demo.Model.ReservationStatus;
 import isa2.demo.Repository.AdventureRepository;
 import isa2.demo.Service.AdventureService;
+import isa2.demo.Service.EntityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class AdventureServiceImpl implements AdventureService {
 
-    @Autowired
-    private AdventureRepository adventureRepository;
+    final
+    AdventureRepository adventureRepository;
+
+    final
+    EntityService entityService;
+
+    public AdventureServiceImpl(AdventureRepository adventureRepository, EntityService entityService) {
+        this.adventureRepository = adventureRepository;
+        this.entityService = entityService;
+    }
 
     @Override
     public List<Adventure> findAll() {
@@ -20,8 +37,8 @@ public class AdventureServiceImpl implements AdventureService {
     }
 
     @Override
-    public Adventure findOne(Integer id) {
-        return this.adventureRepository.findById(id).get();
+    public Optional<Adventure> findOne(Integer id) {
+        return this.adventureRepository.findById(id);
     }
 
     @Override
@@ -30,31 +47,64 @@ public class AdventureServiceImpl implements AdventureService {
     }
 
     @Override
-    public Adventure update(Integer id, Adventure adventure) {
-        Adventure adventureForUpdate = this.adventureRepository.findById(adventure.getId()).get();
-        adventureForUpdate.setName(adventure.getName());
-        adventureForUpdate.setAddress(adventure.getAddress());
-        adventureForUpdate.setCancellationFee(adventure.getCancellationFee());
-        adventureForUpdate.setHouseRules(adventure.getHouseRules());
-        adventureForUpdate.setInstructorBio(adventure.getInstructorBio());
-        adventureForUpdate.setDefaultFishingEquipment(adventure.getDefaultFishingEquipment());
-        adventureForUpdate.setAdditionalServices(adventure.getAdditionalServices());
-        adventureForUpdate.setDescription(adventure.getDescription());
-        adventureForUpdate.setEntityPhoto(adventure.getEntityPhoto());
-        adventureForUpdate.setPhotos(adventure.getPhotos());
-        adventureForUpdate.setPricePerDay(adventure.getPricePerDay());
-        adventureForUpdate.setRentalTimes(adventure.getRentalTimes());
-        adventureForUpdate.setSubscribedClients(adventure.getSubscribedClients());
-        adventureForUpdate.setAdditionalServices(adventure.getAdditionalServices());
-        adventureForUpdate.setReservations(adventure.getReservations());
-        adventureForUpdate.setAverageGrade(adventure.getAverageGrade());
-        adventureForUpdate.setMaxNumberOfGuests(adventure.getMaxNumberOfGuests());
+    public Adventure update(Adventure adventure) {
+        Collection<Reservation> reservations = new ArrayList<>(adventure.getReservations());
+        reservations.removeIf(reservation -> (reservation.getReservationStatus() == ReservationStatus.FREE));
+        if(reservations.isEmpty())
+            return this.adventureRepository.save(adventure);
+        else
+            throw new UnsupportedOperationException("Entity with active reservations can't be deleted");
 
-        return this.adventureRepository.save(adventureForUpdate);
+
     }
 
     @Override
     public void delete(Integer id) {
-        this.adventureRepository.deleteById(id);
+        Adventure adventure = this.adventureRepository.findById(id).orElse(null);
+        Collection<Reservation> reservations = new ArrayList<>(adventure.getReservations());
+        reservations.removeIf(reservation -> (reservation.getReservationStatus() == ReservationStatus.FREE));
+        if(reservations.isEmpty())
+            this.adventureRepository.deleteById(id);
+        else
+            throw new UnsupportedOperationException("Entity with active reservations can't be deleted");
+    }
+
+    @Override
+    public List<Adventure> findAdventuresByInstructor(Owner owner) {
+        return this.adventureRepository.findAllByOwner(owner);
+    }
+
+    @Override
+    public Collection<Adventure> findFreeAdventures(FreeEntityDTO request) throws InvalidInputException {
+        if (request.getStartDate().isAfter(request.getEndDate()) || request.getStartDate().isEqual(request.getEndDate()))
+            throw new InvalidInputException("Invalid start and end date");
+        Collection<Adventure> adventures = adventureRepository.findAll();
+        Collection<Adventure> freeAdventures = new ArrayList<Adventure>();
+        if (request.getGrade() != null && request.getGrade() < 0)
+            throw new InvalidInputException("Grade needs to be positive");
+        if (request.getNumberOfGuests() != null && request.getNumberOfGuests() < 1)
+            throw new InvalidInputException("Number of guests needs to be at least 1");
+        for (Adventure adventure : adventures) {
+            if ((request.getNumberOfGuests() != null && adventure.getMaxNumberOfGuests() < request.getNumberOfGuests()) || (adventure.getAverageGrade() == null && request.getGrade() != null)  || (request.getGrade() != null && adventure.getAverageGrade() < request.getGrade()))
+                break;
+            if (request.getCountry() != null && !request.getCountry().equals(""))
+                if (!request.getCountry().equals(adventure.getAddress().getCountry()))
+                    break;
+            if (request.getCity() != null && !request.getCity().equals(""))
+                if (!request.getCity().equals(adventure.getAddress().getCity()))
+                    break;
+            if (!entityService.isPeriodInRentalTime(adventure, request.getStartDate(), request.getEndDate()))
+                break;
+            else
+                freeAdventures.add(adventure);
+            Collection<Reservation> reservations = adventure.getReservations();
+            for (Reservation reservation : reservations) {
+                if (entityService.doTimeIntervalsIntersect(request.getStartDate(), request.getEndDate(), reservation.getReservedPeriod().getStartDate(), reservation.getReservedPeriod().getEndDate())) {
+                    freeAdventures.remove(adventure);
+                    break;
+                }
+            }
+        }
+        return freeAdventures;
     }
 }
